@@ -1,9 +1,30 @@
-import 'dart:typed_data';
+import 'dart:async';
 
-import 'package:flutter_auto_form/src/models/form.dart';
+import 'package:date_field/date_field.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter_auto_form/flutter_auto_form.dart';
+import 'package:flutter_auto_form/src/configuration/typedef.dart';
 import 'package:flutter_auto_form/src/models/validators/validator.dart';
+import 'package:flutter_auto_form/src/widgets/fields/date_field.dart';
+import 'package:flutter_auto_form/src/widgets/fields/fields.dart';
+import 'package:flutter_auto_form/src/widgets/fields/multiple_sub_form_field.dart';
+import 'package:flutter_auto_form/src/widgets/fields/search_field.dart';
+import 'package:flutter_auto_form/src/widgets/fields/sub_form_field.dart';
 
 import 'field.dart';
+
+/// The different types of field for the [AFTextField].
+enum AFTextFieldType {
+  TEXT,
+  NUMBER,
+  PASSWORD,
+  EMAIL,
+  USERNAME,
+  NEW_PASSWORD,
+  NEW_USERNAME,
+  COLOR,
+  URL
+}
 
 /// The default [Field] extended class used to represent a form's text field.
 class AFTextField<T extends Object> extends Field<T> {
@@ -12,14 +33,18 @@ class AFTextField<T extends Object> extends Field<T> {
     required String name,
     required List<Validator<T>> validators,
     required this.type,
+    this.maxLines = 1,
     T? value,
-  }) : super(id, name, validators) {
+  }) : super(id: id, name: name, validators: validators) {
     super.value = value;
   }
 
   /// Depending on the value of this field, the widget's related instance will act differently.
   ///(For instance, if you choose the password option it will display a hide/display icon).
   final AFTextFieldType type;
+
+  /// {@macro flutter.widgets.editableText.maxLines}
+  final int? maxLines;
 
   @override
   T? parser(T? unparsedValue) {
@@ -29,6 +54,9 @@ class AFTextField<T extends Object> extends Field<T> {
       return unparsedValue;
     }
   }
+
+  @override
+  final FieldWidgetConstructor widgetBuilder = AFTextFieldWidget.new;
 }
 
 /// The default [Field] extended class used to represent a form's number field.
@@ -64,6 +92,62 @@ class AFNumberField<T extends num> extends AFTextField<T> {
   }
 }
 
+/// The default [Field] extended class used to represent a form's number field.
+class AFDateField extends Field<DateTime> {
+  AFDateField({
+    required String id,
+    required String name,
+    required List<Validator<DateTime>> validators,
+  }) : super(
+          id: id,
+          name: name,
+          validators: validators,
+        );
+
+  @override
+  FieldWidgetConstructor get widgetBuilder => ({
+        Key? key,
+        required FieldContext fieldContext,
+      }) =>
+          DateFieldWidget(
+            fieldContext: fieldContext,
+            mode: DateTimeFieldPickerMode.date,
+          );
+
+  @override
+  DateTime? parser(covariant Object? unparsedValue) {
+    return unparsedValue as DateTime?;
+  }
+}
+
+/// The default [Field] extended class used to represent a form's number field.
+class AFTimeField extends Field<DateTime> {
+  AFTimeField({
+    required String id,
+    required String name,
+    required List<Validator<DateTime>> validators,
+  }) : super(
+          id: id,
+          name: name,
+          validators: validators,
+        );
+
+  @override
+  FieldWidgetConstructor get widgetBuilder => ({
+        Key? key,
+        required FieldContext fieldContext,
+      }) =>
+          DateFieldWidget(
+            fieldContext: fieldContext,
+            mode: DateTimeFieldPickerMode.time,
+          );
+
+  @override
+  DateTime? parser(covariant Object? unparsedValue) {
+    return unparsedValue as DateTime?;
+  }
+}
+
 class AFSelectField<T extends Object> extends Field<T> {
   AFSelectField({
     required String id,
@@ -73,9 +157,9 @@ class AFSelectField<T extends Object> extends Field<T> {
     required this.textBuilder,
     required T value,
   }) : super(
-          id,
-          name,
-          validators,
+          id: id,
+          name: name,
+          validators: validators,
         ) {
     super.value = value;
   }
@@ -90,67 +174,9 @@ class AFSelectField<T extends Object> extends Field<T> {
 
     return unparsedValue;
   }
-}
-
-class SimpleFile {
-  SimpleFile(this.name, this.bytes);
-
-  final String name;
-  final Uint8List? bytes;
-}
-
-class AFFileField extends Field<SimpleFile> {
-  AFFileField({
-    required String id,
-    required String name,
-    required List<Validator<SimpleFile>> validators,
-    SimpleFile? value,
-  }) : super(id, name, validators) {
-    super.value = value;
-  }
-
-  String? fileName;
 
   @override
-  SimpleFile? parser(SimpleFile? unparsedValue) => unparsedValue;
-}
-
-class AFSearchModelField<T extends Object> extends Field<T> {
-  AFSearchModelField({
-    required String id,
-    required String name,
-    required List<Validator<Object?>> validators,
-    required this.search,
-  }) : super(
-          id,
-          name,
-          validators,
-        );
-
-  final Future<List<T>> Function(String? query) search;
-
-  @override
-  T? parser(T? unparsedValue) => unparsedValue;
-}
-
-class AFSearchMultipleModelsField<T extends Object> extends Field<List<T>> {
-  AFSearchMultipleModelsField({
-    required String id,
-    required String name,
-    required List<Validator<List<Object?>>> validators,
-    required this.search,
-  }) : super(
-          id,
-          name,
-          validators,
-        );
-
-  final Future<List<T>> Function(String? query) search;
-
-  @override
-  List<T>? parser(covariant List<Object>? unparsedValue) {
-    return unparsedValue?.cast<T>();
-  }
+  final FieldWidgetConstructor widgetBuilder = SelectFieldWidget.new;
 }
 
 class AFBooleanField extends Field<bool> {
@@ -160,77 +186,142 @@ class AFBooleanField extends Field<bool> {
     required List<Validator<bool?>> validators,
     bool value = false,
   }) : super(
-          id,
-          name,
-          validators,
+          id: id,
+          name: name,
+          validators: validators,
         ) {
     super.value = value;
   }
 
   @override
   bool? parser(bool unparsedValue) => unparsedValue;
+
+  @override
+  final FieldWidgetConstructor widgetBuilder = BooleanFieldWidget.new;
 }
 
-class AFFormField<T extends TemplateForm> extends Field<Map<String, Object?>> {
-  AFFormField({
+class AFSearchModelField<T extends Object> extends Field<T> {
+  AFSearchModelField({
     required String id,
     required String name,
-    required this.formGenerator,
-    this.required = true,
-  }) : super(
-          id,
-          name,
-          [],
-        );
+    required List<Validator<Object?>> validators,
+    required this.search,
+  }) : super(id: id, name: name, validators: validators);
 
-  final bool required;
-
-  final T Function() formGenerator;
-
-  T? form;
+  final Future<List<T>> Function(String? query) search;
 
   @override
-  Map<String, Object?>? get value => form?.toMap();
+  T? parser(T? unparsedValue) => unparsedValue;
 
   @override
-  Map<String, Object?>? parser(Map<String, Object?>? unparsedValue) =>
-      unparsedValue;
+  FieldWidgetConstructor get widgetBuilder => SearchModelFieldWidget<T>.new;
 }
 
-class AFMultipleFormField<T extends TemplateForm>
-    extends Field<List<Map<String, Object?>>> {
-  AFMultipleFormField({
+class AFSearchMultipleModelsField<T extends Object> extends Field<List<T>> {
+  AFSearchMultipleModelsField({
     required String id,
     required String name,
-    required this.formGenerator,
-  }) : super(
-          id,
-          name,
-          [],
-        );
+    required List<Validator<List<T>>> validators,
+    required this.search,
+  }) : super(id: id, name: name, validators: validators);
 
-  final T Function() formGenerator;
-
-  final List<T> forms = [];
+  final Future<List<T>> Function(String? query) search;
 
   @override
-  List<Map<String, Object?>> get value => forms.map((e) => e.toMap()).toList();
+  List<T>? get value => super.value;
 
   @override
-  List<Map<String, Object?>> parser(List<Map<String, Object?>>? unparsedValue) {
-    return unparsedValue ?? [];
+  List<T>? parser(covariant List? unparsedValue) {
+    return unparsedValue?.cast<T>();
+  }
+
+  @override
+  final FieldWidgetConstructor widgetBuilder = SearchMultipleModelsField<T>.new;
+}
+
+mixin WithForceErrorStream<T> {
+  final _controller = StreamController<void>.broadcast();
+
+  Stream<void> get forceErrorStream => _controller.stream;
+
+  void sendForceErrorEvent() {
+    _controller.add(null);
   }
 }
 
-/// The different types of field for the [AFTextField].
-enum AFTextFieldType {
-  TEXT,
-  NUMBER,
-  PASSWORD,
-  EMAIL,
-  USERNAME,
-  NEW_PASSWORD,
-  NEW_USERNAME,
-  COLOR,
-  URL
+class AFSubFormField<T extends TemplateForm> extends Field<Map>
+    with WithForceErrorStream {
+  AFSubFormField(
+      {required String id,
+      required String name,
+      required List<Validator<Object?>> validators,
+      required this.form})
+      : super(id: id, name: name, validators: validators);
+
+  final T form;
+
+  @override
+  final FieldWidgetConstructor widgetBuilder = AFSubFormFieldWidget.new;
+
+  @override
+  String? validator([Object? object]) {
+    sendForceErrorEvent();
+
+    if (!form.isComplete()) {
+      return form.getFirstError();
+    }
+
+    return super.validator(object);
+  }
+
+  @override
+  Map get value => form.toMap();
+
+  @override
+  Map? parser(T unparsedValue) {
+    return value;
+  }
+}
+
+class AFMultipleSubFormField<T extends TemplateForm>
+    extends Field<List<Map<String, dynamic>>> with WithForceErrorStream {
+  AFMultipleSubFormField({
+    required String id,
+    required String name,
+    required this.formBuilder,
+    required this.forms,
+    List<Validator<List<Map<String, dynamic>>>> validators = const [],
+  }) : super(
+          id: id,
+          name: name,
+          validators: validators,
+        );
+
+  final T Function() formBuilder;
+  final List<T> forms;
+
+  @override
+  final FieldWidgetConstructor widgetBuilder = AFMultipleSubFormFieldWidget.new;
+
+  @override
+  String? validator([Object? object]) {
+    sendForceErrorEvent();
+
+    for (final form in forms) {
+      if (!form.isComplete()) {
+        return form.getFirstError();
+      }
+    }
+
+    return super.validator(object);
+  }
+
+  @override
+  List<Map<String, dynamic>>? get value =>
+      [for (final form in forms) form.toMap()];
+
+  @override
+  List<Map<String, dynamic>>? parser(covariant Object? unparsedValue) {
+    return value;
+  }
 }
