@@ -1,3 +1,4 @@
+import 'package:flutter_auto_form/src/configuration/typedef.dart';
 import 'package:flutter_auto_form/src/models/validators/validator.dart';
 
 export 'defaults.dart';
@@ -7,8 +8,8 @@ typedef FieldValueParser<T> = T Function(String value);
 /// This class is the base class for any type of custom fields you would
 /// want to create. See the [AFTextField] widget to learn more on how
 /// to extend it.
-abstract class Field<T extends Object> {
-  Field(this.id, this.name, this.validators);
+abstract class Field<T> {
+  Field({required this.id, required this.name, required this.validators});
 
   /// A unique identifier for the field which will be used to retrieve its data.
   final String id;
@@ -32,7 +33,7 @@ abstract class Field<T extends Object> {
 
   /// This method returns null if the field is valid. Otherwhise it will
   /// return the error's string specified in the validator (see [Validator]).
-  String? validate(T? value) {
+  String? validator([Object? object]) {
     for (final Validator validator in validators) {
       final String? error = validator.validate(value);
 
@@ -42,5 +43,8 @@ abstract class Field<T extends Object> {
     return null;
   }
 
+  /// Parses a value into an instance of T.
   T? parser(covariant Object? unparsedValue);
+
+  FieldWidgetConstructor get widgetBuilder;
 }
