@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 export 'defaults.dart';
 
 /// The [Validator] class is used by the [Field] class to validate its input.
@@ -7,7 +5,7 @@ abstract class Validator<T extends Object?> {
   const Validator();
 
   /// This method should return null if the field is valid, a string corresponding
-  /// to the error otherwhise.
+  /// to the error otherwise.
   String? validate(T value);
 }
 
@@ -28,5 +26,8 @@ abstract class ValidatorWithStaticError<T extends Object?>
     }
   }
 
+  /// Returns true or false depending on the validation process. This method
+  /// is then used by the [ValidatorWithStaticError.validate] method to return
+  /// a static error if the result is false.
   bool innerValidate(T? value);
 }
